requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
pymysql
python-dotenv
alembic
cryptography
pytest
